+++ qualys-web-application-scanner/connector.py
@@ -1,5 +1,10 @@
-from connectors.core.connector import Connector, get_logger, ConnectorError
+""" Copyright start
+  Copyright (C) 2008 - 2023 Fortinet Inc.
+  All rights reserved.
+  FORTINET CONFIDENTIAL & FORTINET PROPRIETARY SOURCE CODE
+  Copyright end """
 
+from connectors.core.connector import Connector, get_logger, ConnectorError
 from .operations import operations, _check_health
 
 logger = get_logger('qualys-was')